feat(kotlin): add search helpers `searchForHits` and `searchForFacets`
